Minor refactor to separate user commands
