+++ src/chat/src/commands/index.ts
@@ -3,5 +3,6 @@ import * as BasicCommands from './basic-commands';
 import * as CandleCommands from './candle/candle-commands';
 import * as StreamCommands from './stream-commands';
 import * as NoteCommands from './note-commands';
+import * as UserCommands from './user-commands';
 
-export { AVCommands, BasicCommands, CandleCommands, NoteCommands, StreamCommands };
+export { AVCommands, BasicCommands, CandleCommands, NoteCommands, StreamCommands, UserCommands };
