+++ src/chat/src/twitch-chat.ts
@@ -20,7 +20,8 @@ import {
   BasicCommands,
   CandleCommands,
   NoteCommands,
-  StreamCommands
+  StreamCommands,
+  UserCommands
 } from './commands';
 
 const htmlSanitizeOpts = {
@@ -148,7 +149,7 @@ export class TwitchChat {
     self: boolean
   ) => {
     // Identify user and add to user state if needed
-    await this.saveUser(username);
+    await this.updateAndGetUser(username);
 
     log('info', `${username} has JOINED the channel`);
     const userJoinedEventArg: IUserJoinedEventArg = {
@@ -356,8 +357,6 @@ export class TwitchChat {
       userInfo
     };
 
-    this.emitMessage(SocketIOEvents.OnChatMessage, chatMessageArg);
-
     if (this.activeStream && user.mod) {
       const userEvent: IUserEventArg = {
         streamId: this.activeStream.id,
@@ -368,7 +367,23 @@ export class TwitchChat {
 
     let handledByCommand: boolean = false;
 
-    // Process any commands
+    //Process user commands first before emitting message to hub
+    //so if the user updates, the update is handled before notification
+    for (const userCommand of Object.values(UserCommands)) {
+      handledByCommand = await userCommand(
+        originalMessage,
+        user,
+        this.sendChatMessage
+      );
+      if (handledByCommand) {
+        this.emitMessage(SocketIOEvents.OnChatMessage, chatMessageArg);
+        break;
+      }
+    }
+
+    //Go ahead and emit message to hub before processing the rest of the commands
+    this.emitMessage(SocketIOEvents.OnChatMessage, chatMessageArg);
+
     for (const basicCommand of Object.values(BasicCommands)) {
       handledByCommand = await basicCommand(
         originalMessage,
@@ -501,7 +516,7 @@ export class TwitchChat {
     });
   };
 
-  private saveUser = async (username: string): Promise<IUserInfo> => {
+  private updateAndGetUser = async (username: string): Promise<IUserInfo> => {
     const url = `http://user/update/${username}/false`;
 
     return await get(url).then((user: any) => {
